more.

+++ linuxkm/lkcapi_rsa_glue.c
@@ -607,8 +607,8 @@ static int linuxkm_test_pkcs1_driver(const char * driver, int nbits)
 
     memset(dec, 0, encrypt_len + 1);
     ret = wc_RsaSSL_Verify(sig, encrypt_len, dec, enc_len, key);
-    if (ret <= 0 || ret != enc_len) {
-        pr_err("error: wc_RsaSSL_Verify returned %d, expected %zu\n" , ret,
+    if (ret <= 0 || ret != (int) enc_len) {
+        pr_err("error: wc_RsaSSL_Verify returned %d, expected %d\n" , ret,
                enc_len);
         goto test_pkcs1_end;
     }
